♻️ split core back into multiple fs files
so that it can grow
