+++ packages/core/package.json
@@ -7,7 +7,8 @@
   "type": "module",
   "main": "src/index.ts",
   "scripts": {
-    "build": "tsc && rollup -c",
+    "build": "tsc",
+    "clean": "rimraf tsconfig.tsbuildinfo dist",
     "eslint": "eslint src --ext .ts",
     "eslint:fix": "eslint src --ext .ts --fix",
     "test": "LOG_LEVEL=warn jest --no-watchman",
@@ -19,7 +20,7 @@
     "@babel/preset-typescript": "^7.18.6",
     "@types/jest": "^29.2.3",
     "jest": "^29.3.1",
-    "rollup": "^3.7.2",
+    "rimraf": "3.0.2",
     "typescript": "^4.9.3"
   }
 }
