--- packages/core/rollup.config.js
@@ -1,8 +1,0 @@
-export default {
-  input: "dist/index.js",
-  output: {
-    file: "bundle.js",
-    format: "umd",
-    name: "core",
-  },
-};
