+++ packages/graph-standalone/src/index.html
@@ -84,9 +84,7 @@
           graph.defaultConfiguration({ viewWidth: 400, viewHeight: 400 }),
           svg
         );
-        console.log("loaded graph");
       };
-      console.log("loading page");
     </script>
   </body>
 </html>
