+++ packages/graph/package.json
@@ -8,6 +8,7 @@
   "main": "src/index.ts",
   "scripts": {
     "build": "rollup -c --bundleConfigAsCjs",
+    "clean": "rimraf tsconfig.tsbuildinfo bundle.js dist .rollup.cache",
     "eslint": "eslint src --ext .ts",
     "eslint:fix": "eslint src --ext .ts --fix",
     "test": "LOG_LEVEL=warn jest --no-watchman",
@@ -31,6 +32,7 @@
     "@types/d3-quadtree": "^3.0.2",
     "@types/jest": "^29.2.3",
     "jest": "^29.3.1",
+    "rimraf": "3.0.2",
     "rollup": "^3.7.2",
     "typescript": "^4.9.3"
   }
